requirements.txt
absl-py==1.4.0
ansible==9.3.0
ansible-core==2.16.4
ansible-runner==2.4.0
lockfile==0.12.2
packaging==24.2
pexpect==4.9.0
ptyprocess==0.7.0
python-daemon==3.1.0
PyYAML==6.0.2
